feat: add & save buttons for type management
Added an ADD FIELD button that appears at the end of the list of
fields for content types, and a SAVE button which appears at the end
of the list of properties for each field. ADD FIELD button triggers
a dialog containing a form that takes values for the properties of
the field being created; and upon submission, the field is added to
the database. Similarly, the SAVE button is now responsible for the
database update after some properties of a field is changed.

issues:
    -   Clicking the ADD FIELD button on any content type triggers the
    form for the very last content type in the list.
    -   Modification of the isMandatory and isUnique boolean properties
    is not tracked, and therefore will not trigger a database update by
    itself when the SAVE button is clicked; however, they are still
    properly updated in the database if a database update is triggered
    by modification of another property.